+++ database/migrations/2021_07_04_205848_create_boarding_amounts_table.php
@@ -15,9 +15,9 @@ class CreateBoardingAmountsTable extends Migration
     {
         Schema::create('boarding_amounts', function (Blueprint $table) {
             $table->id();
-            $table->decimal('amount_payable', 8, 2);
-            $table->decimal('total_amount', 8, 2);
-            $table->decimal('balance', 8, 2);
+            $table->integer('amount_payable');
+            $table->integer('total_amount');
+            $table->integer('balance');
             $table->tinyInteger('status');
             $table->integer('collect_boarding_fee_id')->unsigned();
             $table->timestamps();
